fix required changes
