--- app/components/user-preference/style.scss
@@ -1,184 +1,0 @@
-& {
-  margin: 10px 10px;
-  color: $sd-text;
-
-  .float-right {
-    float: right;
-  }
-
-  section {
-    border-radius: 5px 5px 0 0;
-    border: 1px solid $sd-separator;
-    margin-bottom: 20px;
-  }
-
-  h3 {
-    border-radius: 5px 5px 0 0;
-    background-color: $sd-separator;
-    margin: 0;
-    color: $sd-text;
-    padding: 3px 10px;
-    font-weight: $weight-bold;
-    font-size: 16px;
-  }
-
-  h4 {
-    font-weight: $weight-bold;
-    margin: 0;
-    padding: 0;
-  }
-
-  p {
-    margin: 0;
-    padding: 0;
-  }
-
-  ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  li {
-    font-size: 14px;
-    padding: 5px 10px;
-    border-bottom: 1px solid $sd-separator;
-
-    &:last-child {
-      border-width: 0;
-    }
-  }
-
-  a {
-    display: inline-block;
-    text-decoration: none;
-    width: 100px;
-    padding: 3px;
-    text-align: center;
-    border-radius: 3px;
-    border: 1px solid transparent;
-  }
-
-  .right {
-    text-align: right;
-  }
-
-  input.display-job-name {
-    width: 45px;
-  }
-}
-
-.text-input {
-  box-sizing: border-box;
-  margin-top: 10px;
-  outline: 0;
-  height: 45px;
-  width: 100%;
-  border-radius: 2px;
-  border: 2px solid $sd-separator;
-  font-size: 16px;
-  padding-left: 10px;
-  color: $sd-text;
-
-  &:focus {
-    border-color: $sd-running;
-  }
-
-  &::-webkit-input-placeholder {
-    color: $sd-text-med;
-  }
-
-  &:-moz-placeholder {
-    color: $sd-text-med;
-  }
-
-  &::-moz-placeholder {
-    color: $sd-text-med;
-  }
-
-  &:-ms-input-placeholder {
-    color: $sd-text-med;
-  }
-
-  &.bad-text-input {
-    border-color: $sd-failure;
-  }
-
-  &.good-text-input {
-    border-color: $sd-success;
-  }
-
-  &.root-dir {
-    width: calc(100% - 15px);
-  }
-}
-.blue-button {
-  @include transition(all, 50ms);
-  border: none;
-  border-radius: 3px;
-  height: 30px;
-  padding: 0 15px;
-  color: $sd-white;
-  font-size: 14px;
-  background: $sd-running;
-  &:focus {
-    outline: none;
-  }
-  &:hover {
-    background: $sd-link;
-  }
-  &:disabled {
-    background: $sd-text-med;
-  }
-  .saving-loading {
-    display: none;
-  }
-  &.saving {
-    .saving-loading {
-      display: block;
-    }
-    .button-label {
-      display: none;
-    }
-  }
-  &:active {
-    &:not(:disabled) {
-      @include scale(0.97);
-    }
-  }
-}
-
-.reset-button {
-  cursor: pointer;
-  border: 1px solid $sd-running;
-  border-radius: 3px;
-  font-size: 14px;
-  height: 30px;
-  padding: 0 10px;
-  color: $sd-running;
-  background: transparent;
-
-  &:hover {
-    color: $sd-link;
-    @include scale(1);
-  }
-  &:disabled {
-    background: $sd-text-med;
-  }
-  .saving-loading {
-    display: none;
-  }
-  &.saving {
-    .saving-loading {
-      display: block;
-    }
-    .button-label {
-      display: none;
-    }
-  }
-  &:active {
-    &:not(:disabled) {
-      @include scale(0.97);
-    }
-  }
-}
