--- app/components/user-preference/template.hbs
@@ -1,41 +1,0 @@
-{{info-message message=errorMessage type="warning" icon="exclamation-triangle"}}
-{{info-message message=successMessage type="success" icon="check"}}
-<div class="col-xs-12 col-md-8">
-  <section class="preference">
-    <ul>
-      <li>
-        <div class="row">
-          <div class="col-xs-10">
-            <h4>Display Name Length</h4>
-            <p>Setup your own preferred job name length to display on the workflow graph (range: 20 - 99)</p>
-          </div>
-          <div class="col-xs-2 right">
-            <input type="number"
-              class="display-job-name"
-              min={{minDisplayLength}}
-              max={{maxDisplayLength}}   
-              placeholder={{minDisplayLength}}
-              value={{desiredJobNameLength}}
-              oninput={{action (mut displayJobNameLength) value="target.value"}}>
-          </div>
-        </div>
-      </li>
-      <li>
-        <div class="row">
-          <div class="col-xs-12 right">
-            <button {{action "resetUserSettings"}} disabled={{isDisabled}} class="reset-button{{if isSaving " saving"}}">
-              {{fa-icon "refresh"}} Reset
-            </button>
-            <button {{action "updateUserSettings"}} disabled={{isDisabled}} class="blue-button{{if isSaving " saving"}}">
-              <div class="saving-loading">
-                Updating settings
-              </div> 
-              <div class="button-label">Save</div>
-            </button>
-            {{#if isSaving}}{{fa-icon "spinner" spin=true}}{{/if}}        
-          </div>
-        </div> 
-      </li>
-    </ul>
-  </section>
-</div>
